+++ tests/testthat/test-make-array.R
@@ -209,6 +209,29 @@ with_mock_crunch({
         )
         expect_equivalent(varDef, expected)
     })
+    
+    test_that("deriveArray with subvariables creates a VariableDefinition", {
+        expect_json_equivalent(
+            deriveArray(
+                subvariables = list(VariableDefinition(ds$gender == "Male", name = "male")), 
+                name = "Gender MR"
+            ),
+            list(
+                name = "Gender MR",
+                derivation = list(
+                    `function` = "array", 
+                    args = list(list(
+                        `function` = "select",
+                        args = list(list(
+                            map = list(
+                                c(zcl(ds$gender == "Male"), references = list(name = "male"))
+                            )
+                        ), list(value = I("1")))
+                    ))
+                )
+            )
+        )
+    })
 })
 
 with_test_authentication({
@@ -301,6 +324,21 @@ with_test_authentication({
             ds$mr_5 <- makeMRFromText(ds$delim, delim = "; ", name = "myMR")
             expect_true(is.derived(ds$mr_5))
             expect_equivalent(dim(as.vector(ds$mr_5)), c(nrow(ds), 5))
+        })
+    })
+    
+    whereas("deriveArray with subvariables functions as expected", {
+        ds <- newDataset(mrdf)
+        ds$mrVar <- deriveArray(
+            list(VariableDefinition(ds$v4 == "B", name = "subvar name", alias = "sv_alias")), 
+            name = "MR Variable"
+        )
+        
+        test_that("can make MR with formMR", {
+            expect_identical(names(ds), c(names(mrdf), "mrVar"))
+            expect_true(is.MR(ds$mrVar))
+            expect_true(is.Categorical(ds$mrVar$sv_alias))
+            expect_equal(unclass(as.vector(table(ds$mrVar$sv_alias))), c(2, 2))
         })
     })
 })