+++ NEWS.md
@@ -1,7 +1,8 @@
 # crunch (development version)
+* Can now `deriveArray()` using expressions to create the subvariables.
+* New function `slideCategories()` helps you create overlapping categorical 
+variables (#396).
 * crunch will work with upcoming changes to `stringsAsFactors` defaults (#402).
-
-# crunch 1.26.3
 * Exporting a Deck requires (and now includes) a valid JSON body in its request.
 * Documentation for `newSlide` includes examples of `vizType` settings and other improvements.